Switch to RTC Watchdog and keep watchdog counter
Task watchdog resets are not visible in the reset reason,
whereas RTC watchdog is. So now the code knows whether it was reset
because of a hang (watchdog) and can count the number of times this happens.

The watchdog counter is kept in noinit DRAM, meaning it doesn't wear out the
flash memory with its limited write cycles.