Go to deepsleep after a number of infinite loops/hangs
After 3 watchdog reboots due to infinite loops/hangs,
go to sleep for 18 hours. Then try again.

If the result is another 3 watchdog reboots,
go to sleep for 2*18 = 36 hours. Then try again.

If the result is another 3 watchdog reboots,
go to sleep for 3*18 = 54 hours. Then try again, and so on.

This should be a good balance between retrying regularly,
and sleeping plenty, so as not to drain the battery.