Handle wifi status callbacks and write troubleshooting to display
Parse both high level and detailed info about wifi events.

Wrong password, access point not found, usually translate into a disconnect event.
So for those, we print the detailed reason to the display as well as the console.

Also, for typical non-recoverable wifi disconnections, set the watchdog more
agressive so it reboots faster, to conserve battery.