Show update notification on display
The notification is subtle because this might appear for
days, weeks or months before the user decides to update,
and we don't want to annoy them.

Also, updates are probably not critical anyway.

The notification is displayed by the version number (v1.9.5)
changing to UP!1.9.5 after some time. The fact that it changes
helps in drawing attention to it, without being annoying.

Also, updates are checked only every 24 hours, as this is not
time critical and this speeds things up and reduces load.